% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeOutcomes.R
\name{analyzeOutcomes}
\alias{analyzeOutcomes}
\title{Analyze Outcomes}
\usage{
analyzeOutcomes(data)
}
\arguments{
\item{data}{Analysis data generated from generateAnalysisData}
}
\description{
Analyze Outcomes
}
