% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateAnalysisData.R
\name{generateAnalysisData}
\alias{generateAnalysisData}
\title{Generate Analysis Data}
\usage{
generateAnalysisData(
  connectionDetails,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable
)
}
\arguments{
\item{connectionDetails}{Connection details for the database}

\item{cdmDatabaseSchema}{CDM database schema}

\item{cohortDatabaseSchema}{Cohort database schema}

\item{cohortTable}{Cohort table name}
}
\description{
Generate Analysis Data
}
