% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCohorts.R
\name{generateCohorts}
\alias{generateCohorts}
\title{Generate Cohorts}
\usage{
generateCohorts(
  connectionDetails,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable
)
}
\arguments{
\item{connectionDetails}{Connection details for the database}

\item{cdmDatabaseSchema}{CDM database schema}

\item{cohortDatabaseSchema}{Cohort database schema}

\item{cohortTable}{Cohort table name}
}
\description{
Generate Cohorts
}
